% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_stopword.R
\name{add_stopword}
\alias{add_stopword}
\title{Add new stopwords to dutch_stopwords_extra}
\usage{
add_stopword(...)
}
\arguments{
\item{...}{One or more words to add as stopwords.}
}
\value{
A character vector with updated stopwords (`dutch_stopwords_extra`).
}
\description{
This function adds one or more new stopwords to the existing list of Dutch stopwords.
If a word already exists, a message will notify the user.
}
\examples{
add_stopword("er", "bij", "over")
}
