% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_soep_texts}
\alias{example_soep_texts}
\title{Example SOEP Texts}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 500 rows and 5 columns.
}
\usage{
example_soep_texts
}
\description{
Example SOEP Texts
}
\keyword{datasets}
