% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_text.R
\name{preprocess_text}
\alias{preprocess_text}
\title{Preprocess text column}
\usage{
preprocess_text(
  df,
  stopword_list = dutch_stopwords,
  remove_stopwords = TRUE,
  remove_punctuation = TRUE,
  remove_digits = TRUE,
  tolower = TRUE,
  output = c("preprocessed", "tokenised")
)
}
\arguments{
\item{df}{A dataframe with `id` and `text` columns.}

\item{stopword_list}{A character vector of stopwords (default = dutch_stopwords).}

\item{remove_stopwords}{Logical, whether to remove stopwords. Default TRUE.}

\item{remove_punctuation}{Logical, whether to remove standalone punctuation. Default TRUE.}

\item{remove_digits}{Logical, whether to remove standalone digits. Default TRUE.}

\item{tolower}{Logical, whether to force lowercase before tokenisation. Default TRUE.}

\item{output}{Character, "tokenised" to return a tidy df with separate word counts (i.e., tokenised), "preprocessed" to return df with SOEP_clean column (i.e., the complete SOEP text, but cleaned). Default is "preprocessd".}
}
\value{
A dataframe.
}
\description{
Preprocesses text by optionally lowercasing and removing stopwords, punctuation, and/or digits.
}
\examples{
preprocess_text(example_soep_texts)
}
